refactored idle timer prompt to its own component
